:hammer: Add contact preferences migration
