+++ .github/workflows/publish-app-store.yml
@@ -53,7 +53,7 @@ jobs:
 
     - name: Enable App Transport Security
       run: |
-        scripts/prepare-release-ios
+        scripts/prepare-release-ios.sh
 
     - name: Build using Xcode
       uses: yukiarrr/ios-build-action@v1.5.0
