+++ lib/utils/constants/app.ts
@@ -1,74 +1,3 @@
-import { vh, vw } from 'react-native-expo-viewport-units'
-
-// UI
-export const DEVICE_HEIGHT: number = vh(100)
-export const DEVICE_WIDTH: number = vw(100)
-export const POST_PHOTO_SIZE_HEIGHT = 1000
-export const POST_PHOTO_SIZE_WIDTH = 1000
-export const RECOMMENDED_MINIMUM_TAPPABLE_SIZE = 44
-export const STRUCTURE_GO_BACK_BUTTON_SIZE = 24
-export const STRUCTURE_SHARE_BUTTON_SIZE = 24
-export const STRUCTURE_LITTEN_POST_IMAGE_TO_CONTENT_RATIO = 0.55
-export const STRUCTURE_PADDING_MULTIPLIER = 1.2
-export const STRUCTURE_TAB_NAV_ACTIVE_INDICATOR_SIZE = 6
-export const STRUCTURE_TAB_NAV_BORDER_RADIUS = 40
-export const STRUCTURE_TAB_NAV_HEIGHT = 70
-export const STRUCTURE_PADDING_EXTRA =
-  STRUCTURE_TAB_NAV_HEIGHT * (STRUCTURE_PADDING_MULTIPLIER - 1)
-export const STRUCTURE_TAB_NAV_ICON_INDICATOR_DISTANCE = 8
-export const STRUCTURE_TAB_NAV_ICON_SIZE = 28
-export const STRUCTURE_TEMPLATE_SCREEN_BORDER_RADIUS = 26
-export const STRUCTURE_TEMPLATE_SCREEN_HEADER_HEIGHT: number = vh(
-  (1 / 10) * 100,
-)
-export const STRUCTURE_TEMPLATE_SCREEN_HEADER_PADDING_BOTTOM = 24
-export const STRUCTURE_TEMPLATE_SCREEN_HEADER_PROFILE_WIDTH: number = vw(85)
-export const STRUCTURE_TEMPLATE_SCREEN_HEADER_WIDTH: number = vw(90)
-export const STRUCTURE_TEMPLATE_SCREEN_HEADER_MARGIN =
-  (DEVICE_WIDTH - STRUCTURE_TEMPLATE_SCREEN_HEADER_WIDTH) / 2
-export const STRUCTURE_TEMPLATE_SCREEN_PADDING = 24
-export const UI_ABOUT_LOGO_HEIGHT: number = vh(5)
-export const UI_ABOUT_LOGO_RATIO = 3.625
-export const UI_ABOUT_LOGO_WIDTH: number = vw(100 * UI_ABOUT_LOGO_RATIO)
-export const UI_BADGE_ELEMENT_OFFSET = 15
-export const UI_BADGE_SIZE = 20
-export const UI_BADGE_PADDING = UI_BADGE_SIZE / 4
-export const UI_BOTTOM_LOADER_HEIGHT = 40
-export const UI_BUTTON_BORDER_RADIUS = 24
-export const UI_BUTTON_FIXED_WIDTH: number = vw(77)
-export const UI_DROPDOWN_MARGIN = 6
-export const UI_ELEMENT_BORDER_MARGIN = 8
-export const UI_ELEMENT_BORDER_RADIUS = 6
-export const UI_ELEMENT_LIST_HEIGHT = 64
-export const UI_EMPTY_PLACEHOLDER_IMAGE: number = vw(30)
-export const UI_EXTRA_OPTION_SIZE = 20
-export const UI_HIDDEN_OPTION_WIDTH = 100
-export const UI_ICON_SIZE_MEDIUM = 64
-export const UI_ICON_SIZE_MICRO = 12
-export const UI_ICON_SIZE_MINI = 24
-export const UI_ICON_SIZE_SMALL = 40
-export const UI_IMAGE_PLACEHOLDER_ITEM_MARGIN: number = vw(0.5)
-export const UI_IMAGE_PLACEHOLDER_ITEM_SIZE: number = vw(21)
-export const UI_LITTEN_CARD_BORDER_RADIUS = 20
-export const UI_LITTEN_CARD_CONTENT_TO_CONTAINER_RATIO = 0.75
-export const UI_LITTEN_CARD_HEIGHT = 240
-export const UI_LITTEN_CARD_IMAGE_TO_CONTENT_RATIO = 0.47
-export const UI_LITTEN_CARD_SPACING = 28
-export const UI_MAP_BORDER_RADIUS = 8
-export const UI_MAP_MIN_HEIGHT: number = vh(55)
-export const UI_MESSAGE_MIN_INPUT_TOOLBAR_HEIGHT = 54
-export const UI_MODAL_MIN_HEIGHT = 215
-export const UI_OPERATION_STATUS_ICON_SIZE = 80
-export const UI_PRESSED_OPACITY = 0.6
-export const UI_SCREEN_NOAUTH_CTA_HEIGHT = 24
-export const UI_SELECT_OPTION_HEIGHT = 60
-export const USER_AVATAR_SIZE_LARGE = 200
-export const USER_AVATAR_SIZE_MEDIUM = 82
-export const USER_AVATAR_SIZE_MINI = 36
-
-// Screens
-export * from './routes'
-
 // WebViews
 export const WEB_APP_HELP_AND_CONTACT = '/faq'
 export const WEB_APP_TERMS_AND_CONDITIONS = '/terms-and-conditions'
