+++ scripts/setup-emulator.js
