+++ scripts/validate-required-files.sh
