+++ scripts/validate-translations.js
