+++ types/react-native-expo-viewport-units.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-native-expo-viewport-units'
